# public/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EduX Learning Platform API",
    "description": "REST API for the EduX online learning platform. Provides endpoints for user authentication, course management, enrollment, exams, and more.",
    "version": "1.0.0",
    "contact": {
      "name": "EduX Support",
      "email": "support@edux.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    },
    {
      "url": "https://edux.example.com/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/course_suggestion": {
      "post": {
        "tags": ["Courses"],
        "summary": "Search courses",
        "description": "Search for courses by keyword",
        "operationId": "searchCourses",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["search"],
                "properties": {
                  "search": {
                    "type": "string",
                    "description": "Search query"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Search results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Course"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Course": {
        "type": "object",
        "properties": {
          "c_id": {
            "type": "integer",
            "example": 100
          },
          "title": {
            "type": "string",
            "example": "Complete React Course"
          },
          "description": {
            "type": "string",
            "example": "Learn React from scratch"
          },
          "i_id": {
            "type": "integer",
            "example": 1
          },
          "rating": {
            "type": "number",
            "format": "float",
            "example": 4.5
          },
          "student_count": {
            "type": "integer",
            "example": 150
          },
          "wall": {
            "type": "string",
            "example": "/images/react.jpg"
          },
          "field": {
            "type": "string",
            "example": "Web Development"
          },
          "seat": {
            "type": "integer",
            "example": 200
          }
        }
      },
...
    }
  }
}